# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZyNum API - Virtual Numbers Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface RegisterRequest {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  confirmPassword: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  user: UserProfile;
  token: string;
}

export interface Service {
  id: string;
  name: string;
  icon: string;
  category: string;
}

export interface ServicesResponse {
  services: Service[];
}

export interface Country {
  code: string;
  name: string;
  flag: string;
  priceUsd: number;
  priceFcfa: number;
  available: number;
}

export interface CountriesResponse {
  countries: Country[];
}

export type BuyNumberRequestCurrency =
  (typeof BuyNumberRequestCurrency)[keyof typeof BuyNumberRequestCurrency];

export const BuyNumberRequestCurrency = {
  USD: "USD",
  FCFA: "FCFA",
} as const;

export interface BuyNumberRequest {
  service: string;
  country: string;
  currency?: BuyNumberRequestCurrency;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  PENDING: "PENDING",
  RECEIVED: "RECEIVED",
  FINISHED: "FINISHED",
  TIMEOUT: "TIMEOUT",
  BANNED: "BANNED",
  CANCELED: "CANCELED",
} as const;

export interface Order {
  id: string;
  externalId: string;
  phone: string;
  service: string;
  serviceName: string;
  country: string;
  countryName: string;
  status: OrderStatus;
  smsCode?: string | null;
  smsText?: string | null;
  priceUsd: number;
  priceFcfa: number;
  currency: string;
  createdAt: string;
  updatedAt: string;
}

export interface OrderResponse {
  order: Order;
}

export interface SmsStatusResponse {
  order: Order;
}

export interface OrderHistoryResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface BalanceResponse {
  balance: number;
  currency: string;
  lowBalanceThreshold: number;
  isLow: boolean;
}

export interface ApiKeyResponse {
  apiKey: string;
  createdAt: string;
}

export type GetCountriesParams = {
  service?: string;
};

export type GetOrderHistoryParams = {
  page?: number;
  limit?: number;
};